Minor tweaks, comments.
AttachmentOperator now has a default constructor that initializes
includeNonConservativeAttachment to false (code uniformity).

Added comments about booleans that are not re-evaluated properly
in some operators' evaluateXXX members (hasSuperelastics,
includeNonConservativeAttachment, includeNonConservativeIonization).
These are initialized to false once, in the constructor, and possibly
changed to true depending on checks in evaluateXXX. It appears that
these booleans should be set to false at the beginning of these
evaluateXXX members. This may change the results in smart grid runs,
since the values depend on uMax, and a reduction of uMax may result
in some reactions not being considered anymore.