Growth solvers: make ee term part of 'base' matrix
Since the ee terms do not change during the growth matrix solution
procedures, they can be made part of the 'base matrix', whose
diagonals are stored and added to during the iterations. This
avoids the additional task of adding the ee terms (if any)
within the convergence loop. This is also how the MATLAB code
sets up things (using a matrix Mee).