Start discussing the power balance, the coefs mu_u, D_u.
NOTE that the LoKI-B definitions of the energy mobility and diff.
coefficient differ from Hagelaar's definitions by a factor <u>.
The LoKI-B definitions are perhaps better called the thermoelectricity
(beta) and heat diffusion (G), which is what Allis calls them. See
the discussion in the document, and the BOLSIG+ paper (below eq. 62).