Add code coverage (#58)
* Use CodeCoverage.cmake to generate a cobertura coverage report

Generates a `coverage` target that builds the report. Compile with 
`-Dcoverage=ON` to generate this target. Under the hood, `gcovr` is used 
to generate the coverage report.
Registering tests with `target_add_test` will now automatically register the
targets as coverage dependencies.

* Add Codecov coverage workflow

Builds the `coverage` target and uploads the generated `build/coverage.xml`
to Codecov. As this is still a private repository, I have added a GitHub secret
that contains the Codecov token. Codecov will then report the coverage and
the coverage diff compared to the base commit when issuing a pull request.
The additional Codecov checks will fail the workflow when coverage drops.

* Cleanup CMake workflow

* Add Codecov badge to readme