Reworked the non-uniform grid class.
@jophendrikx and @daanboer, FYI. Based on Job's initial implementation,
I completed a possible non-uniform grid implementation. This now
supports updateMaxEnergy and fixes an off-by-one error in the
original implementation (nCells was initialized to nodes.size()).

The idea is that:

 * all constructors delegate to one that accepts the node positions
   (a vector with sorted values in range [0,1]), an initial uMax value
   and a boolean that indicates if the grid is equidistant.
 * the (numCells,uMax) overload sets up an equidistant range and
   calls that function, passing false for isUniform.
 * m_du is set only if isUniform is true, it is zero otherwise.
   Calling du() for a non-uniform mesh results in an exception.
 * m_duCells and m_duNodes are *always* set up
 * the meaning of these vectors has changed. m_duCells contains the
   widths of the cells, m_duNodes contains the 'widths' of the
   nodes (faces). For internal faces, that is the distance between
   the two adjacent cells. For boundary nodes, it is the distance
   from the boundary node to the adjacent internal cell.
   duCells() has the same unumber of elements as getCells(), duNodes()
   the same as getNodes(). There is one du for every cell/node.
 * updateMaxEnergy does a scaling uMaxNew/uMaxOld of m_du and the four
   vectors. Note that that works for uniform as well as non-uniform
   meshes.

As a consequence, an integral over a cell-field can now be calculated
as the inner product of that field and duCells.

A few basic tests were added (tests/test_grid.cpp).

@jophendrikx and @daanboer, please have a look and see if you agree.
This patch (or part of it) can always be reverted, I did not bother
to make yet another branch for this idea. @jophendrikx, it would help
very much if pictures for the boundary cells would also be included.