Support relative path argument to setup file
Propagates base path information to construct paths relative
to the current working directory. This means that the `loki`
executable can now be called with the full relative path to
the desired input file. The restriction that all data has
to reside in the `/input` directory is dropped, and the
nested paths to auxilliary files should simply be written
relative to the main input file directory.