Added a parser for function-call strings.
See class FunctionCall and function makeJsonFromFunctionCall in
Parse.h. This will be used to extract the arguments that are needed
to set up a range, and for a tool that patches a JSON file that
has been converted using the OffSideToJSON code to bring that in
the desired form (see issue #117).