Legacy to JSON conversion (#124)
* Added a parser for function-call strings.

See class FunctionCall and function makeJsonFromFunctionCall in
Parse.h. This will be used to extract the arguments that are needed
to set up a range, and for a tool that patches a JSON file that
has been converted using the OffSideToJSON code to bring that in
the desired form (see issue #117).

* WIP: a tool for converting and patching legacy files.

Add functions legacyToJSON. From the in-code documentation:

/** This function makes a copy of a legacy LoKI-B input file, converted to a
 *  JSON object using offSideToJSON, applies the necessary patches to make the
 *  result a valid LoKI-B JSON input file and returns the result. See section
 *  5 of \cite Manual_2_2_0 for details about the legacy file format.
 *
 *  \todo Document the required, JSON format, provide a reference here.
 *
 *  \sa offSideToJSON
 *  \author Jan van Dijk
 *  \date   30 April 2024
 */

For now, only the workingConditions section is patched. A JSON string in
reducedField like "logspace(-3.0,3.0,7)" is translated in an object that has
named fields such as "min", "max", "nvalues". Also some other working condition
parameters have been updated this way (electronTemperature, electronDensity,
gasPressure, excitationFrequency). "unit" leafs are added (but we still need
to decide how to represent this in the JSON file).

After building, you can type:

  ./ideas/loki_legacytojson ../input/default_lokib_setup.in

Other sections that we (probably) want to different in the JSON input files are
the "X = Y" leafs and the function@parameter leafs in:

    "stateProperties": {
      "energy": [
        "N2(X,v=*) = harmonicOscillatorEnergy",
        "N2(X,v=0,J=*) = rigidRotorEnergy"
      ],
      "population": [
        "N2(X) = 1.0",
        "Nitrogen/N2_vibpop.txt",
        "N2(X,v=0,J=*) = boltzmannPopulation@gasTemperature"
      ],
      "statisticalWeight": [
        "N2(X,v=*) = 1.0",
        "N2(X,v=0,J=*) = rotationalDegeneracy_N2"
      ]
    }

The most important thing is now to decide for the final form of the
JSON files, see issue #117. Since the result of the OffSideToJSON
conversion (without patching) can at present be read by LoKI-B, we
could already decommission the legacy parsers (but we are not in a
hurry, and legacy parsing is still done for the LXCat files).

* Fix key: "X:" -> "X", add gasTemperature parameter.

* Make {OffSide,Legacy}ToJSON code mainstream.

Move this code, including tests, from ideas/ to the main directories
(LoKI-B, source, app, tests). That will allow the usage of the JSON
conversion code in the LoKI-B library and applications.

* Minor fix in comment

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>
Co-authored-by: Daan Boer <daan-boer@hotmail.com>