Allow input conversion to JSON. Render legacy parsing obsolete (but it still works). (#126)
* Allow conversion of legacy and old JSON files.

The main loki app can now be called in two ways:

1.  loki <inputfile>

<inputfile> must be a legacy (.in) file, or a *new-style* JSON file.
In case of an .in file, use the legacy method (Setup) for setting
up LoKI-B, otherwise, use the JSON-based construction.

2.  loki --convert <inputfile>

<inputfile> must be a legacy (.in) file, or a *old-style* JSON file.
In this case JSON-based construction will always be used. If the
input file is a .in file, that will be converted to new-style JSON.
If the input file is old-style JSON (the literal translation of the
off-side .in file, like the JSON files at present in git), that will
be patched to become 'new-style'. In particular, values and range strings
in workingConditions are modified to become JSON objects. See the
overload of legacyToJSON that accepts a json object in LegacyToJSON.h.

NOTE: loki ../input/default_lokib_setup.json will not work, since that
file is (still) old-style JSON. Until converted, use --convert.

@daanboer, FYI. The format must still be agreed on, of course.

* Patch the JSON that is used in the tests.

Once the file format has stabilized, patch the input files (and
inline JSON in the Druyvesteyn test), and remove the calls to
legacyToJSON (and the corresponding include).

* Numerical values should be numbers in JSON input

* Support `JSON` to `JSON` conversion in `loki_legacytojson`

`app/loki_legacytojson` is also formatted using `clang-format`.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>
Co-authored-by: Daan Boer <daan-boer@hotmail.com>