Use JSON to set up LoKI-B, allow multiple outputs.
When a '.in' file is passed, that is converted to JSON. When option
--convert is passed, followed by a JSON file, the JSON file will
be patched from legacy JSON format to the format required by LoKI-B.

In addition, the input file now controls what output will be generated.
when, in "output", a leaf 'writeText: true' is found, *or such leaf
is absent*, text files will be written (the legacy output format).
When an element 'writeJSON' is found, and its value is true, (also)
write JSON. At present, the JSON file is only dumped to the console,
at the end of the simulation. No JSON gets written at present, and
support for JSON may not be 100% up-to-date.

The most important side-effect: No loki::Setup object is created,
LoKI-B/Setup.h is no longer included by the application.