Introduce class GasProperties. (#132)
* Introduce class GasProperties.

This manages the content of a LoKI-B gasProperies section. The
data are stores in a JSON object. It is also possible to create
a default (empty) object, then populate that manually. In addition,
a test program has been added. You can call it as, for example:

 ./app/loki_gasproperties ../input/default_lokib_setup.in /electronKinetics/gasProperties
 ./app/loki_gasproperties ../input/default_lokib_setup.in /electronKinetics/gasProperties mass
 ./app/loki_gasproperties ../input/default_lokib_setup.in /electronKinetics/gasProperties mass N4

This will print, respectively,

 * all properties of all gases
 * the masses of all gases
 * the mass of N4

This class is meant to be used later on to streamline the construction of gases,
doing the paramater initialization at construction time of the gases, rather than
in a post-construction call to loadGasProperties by the mixture.

* Add GasProperties.cpp

* Remove trailing whitespace.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>