Add interfaces for setting/querying properties, tests.
The electron mass is added by the default constructor and by the
one that reads a gasProperties section --- in the latter case only
if it is not set by the sources in gasProperties.

Predicates has(property), has(property,key) and member function
set(property,key,value) have been added. In addition, two member
functions get(property,key) and get(property,key,alternative,warn)
have been added. The first returns the requested property and throws
an error if not available. The second return the requested property
when that is available, the specified alternative value otherwise.
In the latter case, a warning message is written to the console,
when argument warn is true.