Use GasProperties for Gas construction.
In EedfMixture::EedfMixture, create a local GasProperties object
from the configuration node. Pass this (indirectly) to the constructors
of the Gas and the EedfCollisionDataGas objects that are created
when collisions are added. In those constructors, extract the relevant
data from the GasProperties object, when available, and complain when
required data are missing. In the two test files that create a Gas,
set up a minimal GasProperties object manually.

This makes the construction process much more transparent, and
reduces the risk of failing to initialize data members.