Use JSON internally for configuring states.
The stateProperties handling has been changed. In legacyToJSON,
the previous array elements "<states> = <value|function call>"
are rewritten to JSON objects. For the present structure, you
can inspect the output of

  ./app/loki_legacytojson ../input/default_lokib_setup.in

A small excerpt:

      "population": [
        {
          "states": "N2(X)",
          "value": 1.0
        },
        {
          "file": "Nitrogen/N2_vibpop.txt"
        },
        {
          "function": {
            "arguments": [
              "gasTemperature"
            ],
            "name": "boltzmannPopulation"
          },
          "states": "N2(X,v=0,J=*)"
        }
      ],

The sections for "energy" and "statisticalWeights" have the same structure.
The array elements use a "value" or apply a "function" (possibly with arguments)
to the "states". The element type "file" reads the conent from a file. When
a ".json" file is specified, that must have the same content type as the population
(or energy or statisticalWeight) section. From now on, recursion is possible:
also the file can use function calls as arguments, or specify files. When a
non-JSON file is passed as above), the function readLegacyStateProperiesFile
is used to convert the content to JSON. That uses, internally, patchStateProperty,
the same function that is used when parsing the primary input file's stateProperties
subsections.

If changes to the format are needed, change:

 * patchStateProperty (in LegacyToJSON.cpp)
 * loadStatePropertyEntry (in GasMixture.cpp)

The result of this work is that more parsing tasks are avoided, used only when
legacy input is provided; internall JSON is used in more places. Lexing/parsing
now happens only in the LoKI-B code for:

 * the LXCat files
 * parsing the StateEntry strings