Support HighFive (hdf5) in CMakeLists.txt
See issue #119. This adds option -Dwith-hdf5=[0|1]. When 1, cmake
will look for the HighFive C++ bindings for hdf5, and for hdf5 itself.
When HighFive is not available, Fetch_Content will be used to
pul it into the build tree. (The user must still make sure that
hdf5 (libs + devel. packages) are available on the system.