Use GasProperties for Gas construction. (#138)
* Use GasProperties for Gas construction.

In EedfMixture::EedfMixture, create a local GasProperties object
from the configuration node. Pass this (indirectly) to the constructors
of the Gas and the EedfCollisionDataGas objects that are created
when collisions are added. In those constructors, extract the relevant
data from the GasProperties object, when available, and complain when
required data are missing. In the two test files that create a Gas,
set up a minimal GasProperties object manually.

This makes the construction process much more transparent, and
reduces the risk of failing to initialize data members.

* Improved comments, doxygen markup fixes.

Target lokib_doxygen can now be built cleanly again.

* Constness fixes, enabled by early gas initialization

Now that gases are configured at construction time, the EedfMixture
does not need to expose non-const access to the gas mixture to the
outside world anymore. In the gas mixture, a const overload of findGas
was added, and checkGasFractions() and checkPopulations() have been
marked 'const'.

* Add rejection test for argument 1a.

* Include GasProperties.h. No need to include Parse.h.

* Use JSON internally for configuring states.

The stateProperties handling has been changed. In legacyToJSON,
the previous array elements "<states> = <value|function call>"
are rewritten to JSON objects. For the present structure, you
can inspect the output of

  ./app/loki_legacytojson ../input/default_lokib_setup.in

A small excerpt:

      "population": [
        {
          "states": "N2(X)",
          "value": 1.0
        },
        {
          "file": "Nitrogen/N2_vibpop.txt"
        },
        {
          "function": {
            "arguments": [
              "gasTemperature"
            ],
            "name": "boltzmannPopulation"
          },
          "states": "N2(X,v=0,J=*)"
        }
      ],

The sections for "energy" and "statisticalWeights" have the same structure.
The array elements use a "value" or apply a "function" (possibly with arguments)
to the "states". The element type "file" reads the conent from a file. When
a ".json" file is specified, that must have the same content type as the population
(or energy or statisticalWeight) section. From now on, recursion is possible:
also the file can use function calls as arguments, or specify files. When a
non-JSON file is passed as above), the function readLegacyStateProperiesFile
is used to convert the content to JSON. That uses, internally, patchStateProperty,
the same function that is used when parsing the primary input file's stateProperties
subsections.

If changes to the format are needed, change:

 * patchStateProperty (in LegacyToJSON.cpp)
 * loadStatePropertyEntry (in GasMixture.cpp)

The result of this work is that more parsing tasks are avoided, used only when
legacy input is provided; internall JSON is used in more places. Lexing/parsing
now happens only in the LoKI-B code for:

 * the LXCat files
 * parsing the StateEntry strings

* Doxygen markup fixes (target lokib_doxygen).

* Include <sstream>

* Fx compilation with OneAPI for MSW.

* Fix typos.

* Simplify the default GasProperties constructor.

The map is empty initially. No need to check, just set the electron mass.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>