FieldOperator::evaluate: also handle CIEff
Also handle the PT-case using the field operator. The only difference
if the calculation of Omega_x (which enters in Omega_PT), which now
also includes a term that depends on CIEff.

NOTE that the field operator must be re-evaluated in the iterative
loop for the PT-case. Should the evaluation of the field term in the
callbacks when a new E/N is installed be there? Or should that callback
only be added when doing SST, or no growth terms are present?