Support scientific notation for threshold energies (legacy input) (#115)
* Support scientific notation for threshold energies

See issue #106. Before this patch, "PARAM.:  E = 3e-1 eV" would
not be matched; the threshold would be ignored. Instead of 3e-1,
one would have to write 0.3. This patch changes the regular
expression to match floats 3e-1 as well. I consider this code
still very fragile, and we should prevent that misspelled
PARAM.: clauses are ignored. After "PARAM.: E = " we should
insist that this is followed by "<double> eV", for example.

* Parse non-whitespace characters, convert later.

Instead of trying to match a valid floating point number, match a
sequence of non-whitespace characters. When found, convert it to a
threshold energy (in eV) using Parse::getValue. That also checks for
trailing characters, so a number like 0.3X will now result in a
runtime_error.

* Make it an error if an LXCat file cannot be opened.

This makes more sense, and is consistent with the MATLAB version.
Before, only a warning was issued.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>