Complete hf support (#143)
* Add redMobilityHF (of type std::complex<double>).

* Make output dependent on Boltzmann/prescribed.

WIP: so far, only the File output has been updated. Note that the
quantity swarmParameters.redMobilityHF is not yet calculated.

* Add \todo's for the calculation of redMobilityHF.

* FieldOperator::evaluate: also handle CIEff

Also handle the PT-case using the field operator. The only difference
if the calculation of Omega_x (which enters in Omega_PT), which now
also includes a term that depends on CIEff.

NOTE that the field operator must be re-evaluated in the iterative
loop for the PT-case. Should the evaluation of the field term in the
callbacks when a new E/N is installed be there? Or should that callback
only be added when doing SST, or no growth terms are present?

* Implement the HF mobility (if omega!=0.0).

The calculation uses Eigen array expressions, but in the expression for the
imaginary part, a cast to a Vector is needed to compile this. @daanboer,
could you take a look at this?

* Also write the reduced energy D,mu.

* Require that omega==0 if spatial growth is used.

(Like the MATLAB code does.)

* Use `cwiseSqrt()` instead of `array().sqrt()`

This alleviates the need for a cast to `loki::Vector`.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>
Co-authored-by: Daan Boer <daan-boer@hotmail.com>