invertMatrix: Optionally avoid double normalization
This is related to issue #7. When LOKIB_AVOID_NORMALIZING_TWICE is
defined to 1 instead of 0, the new code paths will be activated.
These avoid filling the entire first row of the matrix to impose
the normalization condition on the EEDF. Instead, the first equation
is replaced with M(0,0)*f[0] = M(0,0) (so f[0]=1), and normalization
is done afterwards. We set M(0,0) equal to M(1,1) to avoid increasing
the dynamic range of the matrix elements.