invertMatrix: allow showing/using the bandWidth
Added two defines that control reporting and using the bandWidth as calculated
from the Boltzmann matrix:

  // show the bandwidth?
  #define LOKI_DEBUG_BANDWIDTH 0
  // use the bandwidth to decide on Hessenberg?
  #define LOKI_USE_DEBUG_BANDWIDTH 0

Setting LOKI_DEBUG_BANDWIDTH to 1 shows the calculated bandwidth, the
value of inelasticOperator.hasSuperelastics, and which solver is used.
Setting LOKI_USE_DEBUG_BANDWIDTH to 1 will use the calculated bandwidth
to decide on hessenberg (when the lower bandwidth is equal to -1).
Otherwise, Eigen's partialPivLu is used.