Add code for calculating/using the bandwidth of the Boltzmann matrix (#156)
* Bugfix: include <iostream>

* Add function calculateBandwidth

Add function calculateBandwidth. This calculates the lower and upper
bandwidths (signed) of a dense eigen matrix. See the function documentation
for details. Also add tests/test_linalg.cpp. For now, this only contains
tests for calculateBandwidth, more should be added for other functionality
in the LinearAlgebra.h header.

* invertMatrix: allow showing/using the bandWidth

Added two defines that control reporting and using the bandWidth as calculated
from the Boltzmann matrix:

  // show the bandwidth?
  #define LOKI_DEBUG_BANDWIDTH 0
  // use the bandwidth to decide on Hessenberg?
  #define LOKI_USE_DEBUG_BANDWIDTH 0

Setting LOKI_DEBUG_BANDWIDTH to 1 shows the calculated bandwidth, the
value of inelasticOperator.hasSuperelastics, and which solver is used.
Setting LOKI_USE_DEBUG_BANDWIDTH to 1 will use the calculated bandwidth
to decide on hessenberg (when the lower bandwidth is equal to -1).
Otherwise, Eigen's partialPivLu is used.

* Add test_linalg.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>