LinearAlgebra + superElasticThresholds  cleanup (#159)
* Move unused code in LinearAlgeba.* to ideas/

This changeset addresses issue #158. It moves the unused code from
LinearAlgebra.cpp to ideas/HessenbergExtra.cpp, makes the callable functions in
that file non-inline, adds a header for these functions and adds an example. The
latter contains commented out code that was moved from ElectronKinetics.cpp and
can be fixed and enable if we decide to keep this hessenberg version.

Function 'givens' is now in namespace loki::LinAlg::impl and exposed
by LinearAlgebra.h because that is also used by HessenbergExtra.cpp.

This moves unused, undocumented, untested code out of LoKI-B/ and source/
in anticipation of the first public release.

* Remove unused commented out superElasticThresholds

See #158.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>