Use the matrix bandwidth to select a solver.
This simplifies member invertMatrix. It calculates the bandwidth
of the matrix, and uses that to select TDMA, hessenberg or LU.

Note that calculating the bandwidth is O(n^2). This means that at
present the advantages of TDMA, which is O(n), cannot be profited
from. OTOH, various additions of dense matrices are already happening,
so this does not pessimize the code. On the contrary, the hessenberg
case can now reliably be detected. A future optimization lets
the operators keep track of their matrix bandwidths, and the bandwidth
of the resulting is determined by combining them. In the case of
the elastic, field and CAR operators the bandwidths are always (-1,1),
for others they can be dynamically updated as part of the operator
update.

This changeset also avoids the double renormalization; it solves
the matrix for eedf[0]=1 and does the normalization afterwards.
Note that InelasticOperator::hasSuperelasics is no longer used. I kept
it because the related code can be reused for the abovementioned bandwidth
calculations.