ElectronElectronOperator: evaluate gain and loss
See issue #110. This patch implements the calculation of the power
Gain and Loss, in adition to the net power (which should be zero).
It has been verified to give the same result as MATLAB (after in
MATLAB the problem has been patched in which the matrix eeMatrixAuxA
is partly overwritten in the application of detailed balancing.
This problem has been reported already to 'Lisbon'.)

Resolves #110