Move Range classes to separate files, add tests.
Move the declaration of Range to separate file LoKI-B/Range.h, move
the implementation and plugins to source/Range.cpp. Add unit tests
to tests/test_range.cpp.

This class will be needed later on to hold a series of time points
when the time-dependent calculations are going to be implemented.
This patch prepares for that (and adds missing tests).