Develop and test a consistent discretization of muN
Develop a discretization of muN that uses correct expressions of
df/du at boundary points. Add this result to the test. That shows
small but significant improvements when the new discretization is
used. This patch adds new functions calculateMuNCoefs that calculate
coefficients C such that C.dot(eedf) produces muN. See the code
comments for details.

You may try:

  ninja && ./tests/test_trans_coefs  > out.dat
  gnuplot
   > set logscale y
   > plot "out.dat" u 1:(abs($3/3.39848e+25-1)) index 0 w l, "" u 1:(abs($5/3.39848e+25-1)) index 0 w l, "" u 1:(abs($7/3.39848e+25-1)) index 0 w l

which shows the relative error of the calculation for the three
methods. The second line (column 5) is the new discretization.