Use fgPrimeEnergyIntegral, energyIntegral for DN, muN
Use the integration templates for the evaluation of DN, muN and the
energy variants of these functions in both implementations of
evaluateSwarmParameters. The code in solveSpatialGrowthMatrix()
has not yet been updated, since there the evaluation of muN (or
muE, rather), must be made consistent with the calculation of
matrix ionSpatialGrowthU, which requires a little more work.