Implement `treanorGordietsPopulation` property function
Note that this implementation differs in the Matlab implementation,
which seems to contain a bug when `v > vLimit`, as in this case still
the population is computed using the current state properties instead
of the properties of the 'limit' state.

This implementation should follow equation 77 of `v2.2.0` of the LoKI-B
user manual.