Fix the evaluation of mu and D (#168)
* Add notes on the evaluation of mu and D.

* Develop and test a consistent discretization of muN

Develop a discretization of muN that uses correct expressions of
df/du at boundary points. Add this result to the test. That shows
small but significant improvements when the new discretization is
used. This patch adds new functions calculateMuNCoefs that calculate
coefficients C such that C.dot(eedf) produces muN. See the code
comments for details.

You may try:

  ninja && ./tests/test_trans_coefs  > out.dat
  gnuplot
   > set logscale y
   > plot "out.dat" u 1:(abs($3/3.39848e+25-1)) index 0 w l, "" u 1:(abs($5/3.39848e+25-1)) index 0 w l, "" u 1:(abs($7/3.39848e+25-1)) index 0 w l

which shows the relative error of the calculation for the three
methods. The second line (column 5) is the new discretization.

* Add notes about the spatial growth operator.

* Fix index error in calculateMuNCoefs, add tests.

* Do not include <fstream>.

* Introduce energyIntegral, fgPrimeEnergyIntegral

This functions approximate f*g and f*dg/du over [0,u_max] and return
the result. These can be used for calculating the diffusion coefficients
and mobilities, among other things. This also allows the support
for non-uniform meshes to require changes in fewer places.

* Fix energyIntegral for non-uniform mesh. Add tests.

* Improve comments, argument names.

* Use fgPrimeEnergyIntegral, energyIntegral for DN, muN

Use the integration templates for the evaluation of DN, muN and the
energy variants of these functions in both implementations of
evaluateSwarmParameters. The code in solveSpatialGrowthMatrix()
has not yet been updated, since there the evaluation of muN (or
muE, rather), must be made consistent with the calculation of
matrix ionSpatialGrowthU, which requires a little more work.

* Remove unreachable throw statement.

* fgPrimeEnergyIntegral: support non-uniform meshes

Enable the non-uniform tests in tests/test_integrals.cpp.

* Do not include <cstdlib>. Not needed.

* Use energyIntegral in getMeanEnergy, normalizeEEDF

getMeanEnergy can now also be used for non-uniform meshes.

* Use getMeanEnergy, normalizeEEDF.

This avoids tests for the uniformity of the grid in this code.

* eDensGrowth, coefsCI: support non-uniform grids

Use energyIntegral. coefsCI was renamed from integrandCI and uses
duCells() always, avoiding the need for du(), and making this work
for non-uniform meshes.

* spartial growth: make discretization match muN

Ensure that the calculation of muN is complete (handle the boundaries
correctly), and ensure that the result is consisten with the assembly
of the matrix ionSpatialGrowthU.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>