Interpolate face -> cells before adding 1/sqrt(u).
Omega_cs involves a term proportional to 1/sqrt{u}, but the expression
in which this appears is well-behaved, the singularity is removable.
Do the interpolation from faces to cells before adding this term (for
the cells) to the expressions. (The original expression may even be wrong,
the jury is still out.)

Resolves issue #175