Implement the Power classes out-of-line.
No need for this code to be inlined; this is not performance-critical.
Added member Power::clear() to do the zero-initialization. Call this
from the constructor and remove the direct initialization at the point
of declaration. Update comments on the net ee Power (which should be
zero). This fixes two todo's.