Use to_underlying for enum -> underlying type conversion
Use std::to_underlying to convert an enum value to the value of
its underlying type. make this function available in namespace loki.
When this C++23 function is not available, use the alternative
implementation in Enumeration.h.

This allows the casts to be done without reference to the particular
type that is underlying the enum type.