Eedfcollisions cleanup (#180)
* Remove obsolete \bug.

* EedfCollisionDataGas::addCollision: document.

* addCollision simplify implementation

Use a more straightforward implementation of EedfCollisionDataGas::addCollision.
Avoid '(test ? tgt1 : tgt2)' to determine the assignment target, just use an
if statement.

* Forbid duplicate reactions.

Remove the preprocessor code to conditionally allow duplicate reactions.
See issue #22.

* addCollision: return a reference, not a pointer.

The code has been changed a long time ago to return a non-null pointer
(when no exception is thrown). Change the return type to return a
reference to reflect this state of affairs, and remove the non-null
test at the call site.

* Remove NEW_RATE_COEFFICIENT_OUTPUT code.

Just write output the way the MATLAB code does it: write reactions
sorted by target gas.

* EedfCollision: constness, access and name fixes.

 * mark more members as constant
 * make some more types and members private
 * in loadCollisionsClassic, use the same names for the arguments of
   addCollision, to highlight the similarity with loadCollisionsJSON

* Fixed three minor typos.

* Implement the Power classes out-of-line.

No need for this code to be inlined; this is not performance-critical.
Added member Power::clear() to do the zero-initialization. Call this
from the constructor and remove the direct initialization at the point
of declaration. Update comments on the net ee Power (which should be
zero). This fixes two todo's.

* Remove stray TODO:

* More minor code and comment improvements.

 * update comments of evaluateRateCoefficient, eliminated some todos
 * m_ineRateCoeff, m_supRateCoeff: always set these variables in
   evaluateRateCoefficient, to zero if not calculated. Otherwise
   an old value could be kept, instead of 0, when the grid's uMax
   is reduced (smartGrid).
 * m_rateCoefficients*: make private, add and use constant accessors.
 * m_hasCollisions: use a std::array instead of a naked C-style array.

* Use to_underlying for enum -> underlying type conversion

Use std::to_underlying to convert an enum value to the value of
its underlying type. make this function available in namespace loki.
When this C++23 function is not available, use the alternative
implementation in Enumeration.h.

This allows the casts to be done without reference to the particular
type that is underlying the enum type.

* Improved comment, removed obsolete one.

* Fix doxygen generation (\spec -> \mathrm)

Use \mathrm for tyepesetting element names in the bibliography,
since the \spec command from doc/loki_defs.tex is not available
when Doxygen is run.

* Explain isExtra, remove \todo.

* evaluateRateCoefficients: handle all processes.

Do not skip processes with a threshold that exceeds the grid's uMax.

Resolves #183

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>