Initial Support for HDF5 output (#140)
* Support HighFive (hdf5) in CMakeLists.txt

See issue #119. This adds option -Dwith-hdf5=[0|1]. When 1, cmake
will look for the HighFive C++ bindings for hdf5, and for hdf5 itself.
When HighFive is not available, Fetch_Content will be used to
pul it into the build tree. (The user must still make sure that
hdf5 (libs + devel. packages) are available on the system.

* Allow writing (only the EEDF, so far) to an hdf5 file.

When LOKIB_USE_HIGHFIVE, an output producer of type HDF5Output
will be created in loki_main.cpp, depending on the input file.
In input/default_lokib_setup.in I added:

  writeHDF5: true
  HDF5File: output.h5

The HDF5 output task has been implemented only partly: only
the EEDF is written (that is: an energy data set and the (corresponding)
EEDF samples. The data are written to subsections, one for each case
(for the modified input file, to the group ReducedField_50). If
you have hdf5 installed, you may run "h5dump output.h5" to show
the contents. Or "h5ls -d output.h5/ReducedField_50/energy" and
"h5ls -d output.h5/ReducedField_50/eedf" to retrieve the energy
and EEDF values.

This work is best continued after discussing the desired file
structure with the colleagues in Lisbon and Cordoba.

I borrowed some code from @Rickbude to implement this quickly, including
the CMakeLists.txt bits. Credits go to him in (a large) part.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>