Add a discretizer code idea.
That implements the exponential scheme for a system that consists
of multiple convective-diffusive contributions (at present only
elastic and field). The power terms are determined from the
(partial) discretization matrices, that are coupled via the
global (total) Peclet numbers.

These ideas are explained in the additions to the cpp_notes document
that were introduced in f73387d44cdcc79daaedacf44b19be8838bd1f34.