Keep (E/N)^2 out of the fieldOperator tasks.
The field operator's discretize member does not require E/N as
parameter anymore. The factor (E/N)^2 needs to be applied to the
resulting matrix and now needs to passed to evaluatePower(). This
follows the MATLAB code, where this change was made to facilitate
the time-dependent simulations, where we have E(t).

This also allows the calculation of the mobility from the fieldMatrix,
as explained in the cpp_notes.