Add a discretizer code idea. (#202)
* Add a discretizer code idea.

That implements the exponential scheme for a system that consists
of multiple convective-diffusive contributions (at present only
elastic and field). The power terms are determined from the
(partial) discretization matrices, that are coupled via the
global (total) Peclet numbers.

These ideas are explained in the additions to the cpp_notes document
that were introduced in f73387d44cdcc79daaedacf44b19be8838bd1f34.

* gandle overflow for positive P, not for negative P.

* Fix comment.

* Introduce free function solveEEDF.

Moved the implementation of ElectronKineticsBoltzmann::invertMatrix
to new free function solveEEDF(eedf,matrix,grid). Call this function
from ElectronKineticsBoltzmann::invertMatrix.

This allows the usage of this function in other contexts where
M*f=0 must be solved, subject to the condition that f is normalized.

* Various interface improvements.

 * use function solveEEDF
 * make the operators aware of the grid (store a reference)
 * implement a member for calculating the power terms

* Add Amemiya1997, Nam2008.

* Use a ghost cell for the (upper) boundary flux.

Fluxes are now also printed to the console (index 1 in Gnuplot,
index 0 is the eedf).

* Assorted fixes. Discuss the defs. of matrices M.

* Updates on scheme, bound. cond. implementations.

* Merged the various bits on discretization, schemes.

* Minor improvements in the discretization parts.

* Fix, improve the CAR presentation, field ops.

* Minor fixes, add notes on the prescribed EEDF.

* Add notes on the prescribed EEDF function.

 * Discuss properites of f_s(u).
 * Add a figure of sqrt(u)f_s(u) (reduced form).
 * Add reference Gudmundsson2001.
 * Modify .gitignore to 'accept' pdf and gnu(plot) files.

---------

Co-authored-by: Jan van Dijk <jan@epgmod.phys.tue.nl>