Add a new `invertMatrixFast` function
This function solves the system `H f = -A f*`, where A is the superelastic
part of the Boltzmann matrix, and H is the remaining (upper-Hessenberg)
part of the matrix (thus B = H + A). This method requires iteration,
however it allows for the use of the Hessenberg solver instead of a
full LU decomposition. Moreover, usually not many iterations are required
to reach convergence. A speedup of 5x is seen for the N2 input file. The
problem with the current implementation is that it struggles at low E/N.
Perhaps a Maxwellian f0 can solve this problem (currently A is ignored
in the first iteration).