Add a method to easily add unit/regression tests
The added test will perform all tests defined in the `valid-inputs`
directory. A correct test is defined by a folder that contains an input
file `input.in` and an expected output file `output.json`. The test will
run a simulation based on `input.in`, and compare its output to
`output.json`. For now this is a simple equivalency check on the resulting
JSON output object and the expected JSON object defined in `output.json`.